Added physics, improved control panel, cameras
Control panel now has sliders and text boxes for easy and granular input.

Added an improved dynamic camera view so the user can fly along with the
projectile.

Added physics calculations for Lorentz factor, relativistic kinetic energy.

Added rudimentary crater logic for when the projectile does not have enough
energy to destroy Earth. Crater size is hard coded and position is not based
on impact location, needs to be improved.